requirements (1).txt
Flask==2.3.2
python-telegram-bot==13.15
gunicorn==21.2.0
reportlab==4.1.0
requests==2.31.0
psycopg2-binary==2.9.9
pytz==2024.1
python-dotenv==1.0.1
